requirements.txt
discord.py>=2.4,<3.0
python-dotenv>=1.0,<2.0

